refactor: renaming variables,types and functions
