--- src/types/book.ts
@@ -1,6 +1,0 @@
-export type Book = {
-  id: string;
-  title: string;
-  authors: string[];
-  year: string;
-};
